feat: mudando tipo da variavel caloria na classe comida

+++ Generics_Exercises/armazem/Comida.java
@@ -3,7 +3,7 @@ import java.util.Locale;
 
 public class Comida{
     private String nome;
-    private int calorias;
+    private double calorias;
     private double preco;
     static {
         Locale.setDefault(new Locale("pt", "BR"));
@@ -17,11 +17,11 @@ public class Comida{
         this.nome = nome;
     }
 
-    public int getCalorias() {
+    public double getCalorias() {
         return calorias;
     }
 
-    public void setCalorias(int calorias) {
+    public void setCalorias(double calorias) {
         this.calorias = calorias;
     }
 
@@ -37,11 +37,11 @@ public class Comida{
     public String toString() {
 
 
-        return String.format("[%s] %d R$ %.6f", nome, calorias, preco);
+        return String.format("[%s] %.6f R$ %.6f", nome, calorias, preco);
 
     }
 
-    public Comida(String nome, int calorias, double preco) {
+    public Comida(String nome, double calorias, double preco) {
         this.nome = nome;
         this.calorias = calorias;
         this.preco = preco;
